refactor: renamed errors.ts to api-utils.ts
