+++ src/app/api/teams/[teamId]/roblox-credentials/route.ts
@@ -1,6 +1,6 @@
 import { auth } from "@/src/lib/auth";
 import { RobloxCredentialInfoSchema } from "@/src/lib/types/roblox-credentials-types";
-import { AccessDenied, ErrorToNextResponse } from "@/src/lib/utils/errors";
+import { AccessDenied, ErrorToNextResponse } from "@/src/lib/utils/api-utils";
 import { RobloxCredentialsService } from "@/src/services/RobloxCredentialsService";
 import { headers } from "next/headers";
 import { NextResponse } from "next/server";
